Move ProposalCommentVisibility to abstracts module
and rename to AbstractCommentVisibility
